apps/requirements.txt
Flask
Flask-bootstrap4
pylint
Flask-session
